+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissu.java
@@ -37,7 +37,7 @@ public class KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissu extends
      */
     public KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissu(HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(対象年月, UrSystemErrorMessages.値がnull.getReplacedMessage("対象年月"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
@@ -106,7 +106,7 @@ public class KyotakuKeikakuJikoSakuseiTankiNyushoRiyoNissu extends
      *
      * @return 履歴番号
      */
-    public Decimal get履歴番号() {
+    public int get履歴番号() {
         return entity.getRirekiNo();
     }
